requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gTTS==2.5.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
munkres==1.1.4
nest_asyncio==1.6.0
numpy==1.26.4
openai==0.28.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pandas-stubs==2.2.1.240316
parso==0.8.4
pathtools==0.1.2
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
plotly==5.21.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pydeck==0.8.0b4
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.33.0
streamlit-audiorecorder==0.0.4
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
types-pytz==2024.1.0.20240417
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.1
wandb==0.16.5
watchdog==4.0.0
wcwidth==0.2.13
wheel==0.43.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
yarl==1.9.4
zipp==3.17.0
